WIP:  Old outstanding changes.
These changes were made (but not staged) some two years ago but I never got around to completing them.  Committing them now so they aren't lost.

It might be wise to abandon this branch, as it is (as I recall) an obsession with optimizing rhythm complexity calculations assuming an elementary rhythm complexity measure that was chosen primarily because of its mathematical simplicity and not so much on the basis of psychoacoustic principles or experimental comparison.